requirements.txt
mysql-connector-python
cx_Freeze